Pending Transaction service implemented with IPeriodicTask interface
